Improve balance(...) function in Map.scala
As the current implementation does pattern-matching more than once, we can
re-implement the function based on the haskell Map.balance function[1, 2]
so that matches happen at most once.

Suppose we call balance(...) -> rotateL(...) -> singleL(...) of the current implementation,
we do pattern-matching twice for the right subtree r: A ==>> B. This type of redundancy is
removed by this patch.

[1] https://github.com/haskell/containers/blob/v0.5.7.1/Data/Map/Base.hs#L2554-L2585
[2] https://github.com/haskell/containers/commit/6993f3df0dfe0b6d744456cf073bf6b5730b2e52