Add utility function trim to object ==>>(Map)
To improve some functions that use trim, such as union(...), we add
trim so that two boundaries lo and hi are used instead of creating
a new function within union(...) for example.[1,2]

As the new trim function is complicated, its corresponding test was added.
The old trim function is now deprecated and should(will) be removed as
utility functions are not supposed to be used by users.

[1] https://github.com/haskell/containers/commit/185a8bc428a6bc9aaeec638a9c582be6b1e648c0
[2] https://github.com/haskell/containers/blob/v0.5.7.1/Data/Map/Base.hs#L2252-L2270