Add utility function trimLookupLo to object ==>>(Map)
To improve some functions that use trimLookupLo, such as unionWithKey(...),
we add trimLookupLo so that two boundaries lk and hkOption are used
instead of creating a new function within unionWithKey(...) for example.[1,2]

As the new trimLookupLo function is complicated, its corresponding test was added.
The old trimLookupLo function is now deprecated and should(will) be removed as
utility functions are not supposed to be used by users.

[1] https://github.com/haskell/containers/commit/185a8bc428a6bc9aaeec638a9c5
[2] https://github.com/haskell/containers/blob/v0.5.7.1/Data/Map/Base.hs#L2272-L2300